Değişkenlerdeki cast problemi ve her zonede meat dumpling gelme sorunu düzeltilmiştir.

+++ src/GameServer/MagicInstance.cpp
@@ -2408,12 +2408,12 @@ short MagicInstance::GetMagicDamage(Unit *pTarget, int total_hit, int attribute)
 
 		damage = (230 * total_hit / (total_r + 250));
 		random = myrand(0, damage);
-		damage = (random * 0.3f + (damage * 0.85f)) - sMagicAmount;
+		damage = (short)(random * 0.3f + (damage * 0.85f)) - sMagicAmount;
 
 		if (pSkillCaster->isNPC())
 			damage -= ((3 * righthand_damage) + (3 * attribute_damage));
 		else if (attribute != MAGIC_R)	// Only if the staff has an attribute.
-			damage -= (((righthand_damage * 0.8f) + (righthand_damage * pSkillCaster->GetLevel()) / 60) + ((attribute_damage * 0.8f) + (attribute_damage * pSkillCaster->GetLevel()) / 30));
+			damage -= (short)(((righthand_damage * 0.8f) + (righthand_damage * pSkillCaster->GetLevel()) / 60) + ((attribute_damage * 0.8f) + (attribute_damage * pSkillCaster->GetLevel()) / 30));
 		if (pTarget->m_bMagicDamageReduction < 100)
 			damage = damage * pTarget->m_bMagicDamageReduction / 100;
 	}
